Move clickhouseSupport check to check_redis_prefs
