+++ scripts/lua/get_flow_process_tree.lua
@@ -7,9 +7,6 @@ package.path = dirs.installdir .. "/scripts/lua/modules/?.lua;" .. package.path
 
 require "lua_utils"
 local json = require "dkjson"
-require "lua_utils_generic"
-require "label_utils"
-require "lua_utils_gui"
 sendHTTPHeader('text/json')
 
 local flow_key = _GET["flow_key"]
