+++ scripts/lua/modules/alert_utils.lua
@@ -19,8 +19,7 @@ local format_utils = require "format_utils"
 local alerts_api = require "alerts_api"
 local icmp_utils = require "icmp_utils"
 local flow_risk_utils = require "flow_risk_utils"
-require "lua_utils_generic"
-
+require "check_redis_prefs"
 
 local shaper_utils = nil
