+++ scripts/lua/modules/check_redis_prefs.lua
@@ -219,3 +219,30 @@ function getThroughputType()
   
   return throughput_type
 end
+
+-- ##############################################
+
+function hasClickHouseSupport()
+  local auth = require "auth"
+
+  if not (ntop.isPro() or ntop.isnEdgeEnterprise())
+     or ntop.isWindows() then
+     return false
+  end
+
+  -- Don't allow nIndex for unauthorized users
+  if not auth.has_capability(auth.capabilities.historical_flows) then
+     return false
+  end
+
+  -- TODO optimize
+  if prefs == nil then
+     prefs = ntop.getPrefs()
+  end
+
+  if prefs.is_dump_flows_to_clickhouse_enabled then
+     return true
+  end
+
+  return false
+end
