+++ scripts/lua/modules/lua_utils_generic.lua
@@ -529,37 +529,12 @@ function printInterfaceIndex(idx)
    end
 end
 
--- ##############################################
-
-function hasClickHouseSupport()
-   local auth = require "auth"
-
-   if not (ntop.isPro() or ntop.isnEdgeEnterprise())
-      or ntop.isWindows() then
-      return false
-   end
-
-   -- Don't allow nIndex for unauthorized users
-   if not auth.has_capability(auth.capabilities.historical_flows) then
-      return false
-   end
-
-   -- TODO optimize
-   if prefs == nil then
-      prefs = ntop.getPrefs()
-   end
-
-   if prefs.is_dump_flows_to_clickhouse_enabled then
-      return true
-   end
-
-   return false
-end
 
 -- ##############################################
 
 -- NOTE: global nindex support may be enabled but some disable on some interfaces
 function interfaceHasClickHouseSupport()
+   require "check_redis_prefs"
    return(hasClickHouseSupport() and ntop.getPrefs()["is_dump_flows_to_clickhouse_enabled"])
 end
