+++ scripts/lua/rest/v2/get/host/find.lua
@@ -11,8 +11,6 @@ local tag_utils = require "tag_utils"
 local snmp_utils
 local snmp_location
 
-require "lua_utils_generic"
-
 if ntop.isPro() then
    package.path = dirs.installdir .. "/scripts/lua/pro/modules/?.lua;" .. package.path
    snmp_utils = require "snmp_utils"
